backend/requirements.txt
flask==3.1.2
mysql-connector-python==8.0.33
pytest==7.4.0
flask-cors==6.0.1
